+++ lib/orders.d.ts
@@ -1,6 +1,7 @@
 ﻿import * as Request from './request';
 import * as PurchaseOrder from './orders/purchase';
-export { PurchaseOrder };
+import * as Shipment from './orders/shipment';
+export { PurchaseOrder, Shipment };
 /**
  * The request for getting All Released Orders.
  *
@@ -31,6 +32,19 @@ export interface AcknowledgeOrderRequest extends Request.RequestHeaders {
     PurchaseOrderId: number;
 }
 /**
+ * The request for requesting a shipping update.
+ */
+export interface ShippingUpdateRequest extends Request.RequestHeaders {
+    /**
+     * A unique ID associated with the seller's purchase order.
+     */
+    PurchaseOrderId: number;
+    /**
+     * Information about the shipment.
+     */
+    PurchaseOrderShipment: Shipment.OrderShipmentRequest;
+}
+/**
  * Retrieves all the orders with line items that are in the "created" status, that is,
  * these orders have been released from the Walmart Order Management System to the
  * seller for processing. The released orders are the orders that are ready for a
@@ -38,8 +52,8 @@ export interface AcknowledgeOrderRequest extends Request.RequestHeaders {
  *
  * @param params The GetAllReleaseRequest.
  *
- * @returns Result of the Promise is an Orders.PurchaseOrder.PurchaseOrderResponse if
- *          `Accept` is `application/json`.
+ * @return Result of the Promise is an Orders.PurchaseOrder.PurchaseOrderResponse if
+ *         `Accept` is `application/json`.
  *
  * @see {@link https://developer.walmart.com/#/apicenter/marketPlace/latest#getAllReleasedOrders}
  */
@@ -58,9 +72,24 @@ export declare function getAllReleased(params: GetAllReleasedRequest): PromiseLi
  *
  * @param params The AcknowledgeOrderRequest.
  *
- * @returns Result of the Promise is an Orders.PurchaseOrder.PurchaseOrderResponse if
- *          `Accept` is `application/json`.
+ * @return Result of the Promise is an Orders.PurchaseOrder.PurchaseOrderResponse if
+ *         `Accept` is `application/json`.
  *
  * @see {@link https://developer.walmart.com/#/apicenter/marketPlace/latest#acknowledgingOrders}
  */
 export declare function ackOrder(params: AcknowledgeOrderRequest): PromiseLike<any>;
+/**
+ * Updates the status of order lines to "Shipped" and triggers the charge to the
+ * customer. Sellers must acknowledge your orders before sending a shipping update to
+ * avoid underselling. After canceling your order, update the inventory for the
+ * canceled order and send it in the next inventory feed. An order line, once marked
+ * as shipped, cannot be updated.
+ *
+ * @param params The ShippingUpdateRequest.
+ *
+ * @return Result of the Promise is an Orders.PurchaseOrder.PurchaseOrderResponse if
+ *         `Accept` is `application/json`.
+ *
+ * @see {@link https://developer.walmart.com/#/apicenter/marketPlace/latest#shippingNotificationsUpdates}
+ */
+export declare function postShippingUpdate(params: ShippingUpdateRequest): PromiseLike<any>;